PIN-3192 Resolved PR issues
