+++ eservices-monitoring-exporter/src/main/resources/application-standalone.conf
@@ -1,4 +1,4 @@
-interop-be-eservices-file-generator {
+interop-be-eservices-monitoring-exporter {
   storage {
     kind      = "file"
     bucket    = "bucket"
