+++ eservices-monitoring-exporter/src/main/resources/application.conf
@@ -1,9 +1,10 @@
-interop-be-eservices-file-generator {
+interop-be-eservices-monitoring-exporter {
   storage {
     kind      = "S3"
-    bucket    = ${DASHBOARD_DATA_BUCKET}
-    filename  = ${DASHBOARD_DATA_FILENAME}
+    bucket    = ${ESERVICE_MONITORING_EXPORT_BUCKET}
+    filename  = ${ESERVICE_MONITORING_EXPORT_FILENAME}
     asndjson  = "false"
+    asndjson  = {?IS_NDJSON_FORMAT}
   }
   read-model {
     db-name           = ${READ_MODEL_DB_NAME}
