+++ eservices-monitoring-exporter/src/main/resources/logback.xml
@@ -33,6 +33,7 @@
     </appender>
 
     <logger name="it.pagopa.interop" level="INFO" />
+    <logger name="org.mongodb" level="WARN"/>
 
     <root name="application" level="ERROR">
         <appender-ref ref="ASYNC_STDOUT" />
