+++ eservices-monitoring-exporter/src/main/scala/it/pagopa/interop/eservicesmonitoringexporter/model/EServiceDB.scala
@@ -1,11 +1,10 @@
-package it.pagopa.interop.eservicesfilegenerator.model
+package it.pagopa.interop.eservicesmonitoringexporter.model
 
 import spray.json.DefaultJsonProtocol._
 import it.pagopa.interop.catalogmanagement.{model => DependencyCatalog}
 import it.pagopa.interop.catalogmanagement.model.persistence.JsonFormats._
-import it.pagopa.interop.commons.utils.SprayCommonFormats.{offsetDateTimeFormat, uuidFormat}
+import it.pagopa.interop.commons.utils.SprayCommonFormats.uuidFormat
 import spray.json._
-import java.time.OffsetDateTime
 import java.util.UUID
 
 final case class EServiceDB(
@@ -20,11 +19,10 @@ final case class DescriptorDB(
   id: UUID,
   state: DependencyCatalog.CatalogDescriptorState,
   serverUrls: Seq[String],
-  createdAt: OffsetDateTime,
   version: String
 )
 
 object EServiceDB {
-  implicit val descriptorDBFormat: RootJsonFormat[DescriptorDB] = jsonFormat5(DescriptorDB.apply)
+  implicit val descriptorDBFormat: RootJsonFormat[DescriptorDB] = jsonFormat4(DescriptorDB.apply)
   implicit val eServiceDBFormat: RootJsonFormat[EServiceDB]     = jsonFormat5(EServiceDB.apply)
 }
