+++ eservices-monitoring-exporter/src/main/scala/it/pagopa/interop/eservicesmonitoringexporter/util/Jobs.scala
@@ -1,15 +1,15 @@
-package it.pagopa.interop.eservicesfilegenerator
+package it.pagopa.interop.eservicesmonitoringexporter.util
 
 import it.pagopa.interop.commons.cqrs.service.ReadModelService
-import it.pagopa.interop.eservicesfilegenerator.util._
+import it.pagopa.interop.eservicesmonitoringexporter.util._
 import it.pagopa.interop.commons.logging._
 import com.typesafe.scalalogging.LoggerTakingImplicit
 import it.pagopa.interop.commons.files.service.FileManager
 import io.circe.syntax._
-import it.pagopa.interop.eservicesfilegenerator.util.Configuration
-import it.pagopa.interop.eservicesfilegenerator.model.EServiceDB
-import it.pagopa.interop.eservicesfilegenerator.model.EService
-import it.pagopa.interop.eservicesfilegenerator.model.EService._
+import it.pagopa.interop.eservicesmonitoringexporter.util.Configuration
+import it.pagopa.interop.eservicesmonitoringexporter.model.EServiceDB
+import it.pagopa.interop.eservicesmonitoringexporter.model.EService
+import it.pagopa.interop.eservicesmonitoringexporter.model.EService._
 
 import scala.concurrent.{Future, ExecutionContext}
